Computational efficiency, reaktoro plotting updates, and monte plot updates.
Added a PrecomputeEOS() function that precomputes every EOS that will be used in PlanetProfile run.
A majority of PlanetProfile runtime is slowed down by computing EOS for each planet (ocean EOS, ice EOS, mantle/iron core EOS).
While PP tries to save these EOS to a global EOSlist, this EOSlist is not shared among workers in parallel processing and oftentimes is not robust for the ice EOS generation.
Thus, we now precompute all the EOSs for large scale explorations (Exploreogram, Inductogram, Monte Carlo) beforehand and save to EOSlist before it is shared to the individual workers.
This involves generating P-T grids for the ocean and ice EOS that are based on input parameters (deltaP, deltaT, etc.) - see PrecomputeEOS() for more details.
This has dramatically reduced runtime of large scale explorations to a few seconds per model run, depending on properties being calculated (i.e. tidal Love numbers, magnetic induction icnrease runtime).
One note is that the EOSlist object becomes very large (scale of GBs), and this is stored in RAM, which may cause crashes if users' machines have limited RAM.
This is especially pertinent to Windows users, where the EOSlist is copied for each worker in parallel processing due to the way Windows handles multiprocessing, so it is something to be aware of.

We also add updates to reaktoro plotting to plot the volume of solid species, rather than mol/kg.
We also introduce a new monte plot to show scatter by ocean composition coloring. This needs to be improved in future.