Computational efficiency updates/bugfixes, timing update, PyALMA updates, and new thermodynamic plots.
Adds some bug fixes when pre-generating all EOS and allows for parallel processing to pre-compute CustomSolution EOS grids, which is the most time intensive part of the EOS generation.

Adds the TimingStruct and timing logger, that allows the code to be more comprehensively timed. This is a WIP, but is meant to improve debugging for computational effiency.

Sets default of pyalma caclulations to complex and saves the amplitude and phase delay calculations, which can now be plotted in the ExploreOgram exploration plots.

Adds two new thermodynamic plots to PlanetProfile individual runs - PlotMeltingCurves and PlotHydrosphereThermodynamics (toggled by Params.PLOT_MELTING_CURVES and Params.PLOT_HYDROSPHERE_THERMODYNAMICS  respectively). The former plots the melting curve for the ocean EOS used, and the second plots the profile of  thermodynamic properties used (density, alpha, Cp).

Also some changes to the reaction calculation in Reaktoro, where we now include the reaction in the speciation calculation to both improve runtime and also allow equilibrium/disequilibrium amounts of reaction substructs to show in hydrospecies plots. Namely, we no longer use RktAffinity(), so should remove that in the future.